fix CODER_ACCESS_URL helm chart logic
I'm not certain this logic is works for our use case.
We use envFrom to set CODER_ACCESS_URL. But the template
assumes it it is not set via coder.env directly, to set a default.

This ends up being http://coder.coder.svc.cluster.local

Which breaks some things.